requirements.txt
faiss-cpu
numpy
python-dotenv
sqlalchemy>=1.4
psycopg2-binary
cryptography
